wasm-smith: Represent the components index space with `Vec<ComponentType>`
Rather than an indirect reference to the nested component or import.

This requires the ability to get the `ComponentType` for components that we
generate ourselves, which in turn requires that we are able to do ~deep copies
of type definitions into new types index spaces because the component might
import or export functions that reference types that the parent index space does
not have a definition for. We can't use aliases in this scenario because they
only allow an inner component to reuse type defs from an outer component, but
this is the opposite situation. This deep copying is further complicated by the
fact that a component which imports another component (for example) needs to
have a type for that component import, and that component type introduces a new
index space, so deep copying needs to deal with and keep track of many index
spaces at the same time.

Future commits will leverage this component type information for generating
instantiations.